analyzer: Un-nest NuGet package metadata classes
Move classes used to deserialize NuGet package metadata to the top level
and move the formerly surrounding `NuGetAllPackageData` class as a private
`AllPackageData` class to its only user, the `NuGetSupport` class. This
further cleans up the code for readability.
